+++ scripts/docker/portable/publish.sh
@@ -3,7 +3,7 @@
 #############################################################################
 ##
 ##   Logtalk Docker publishing script for stable versions
-##   Last updated on June 7, 2024
+##   Last updated on July 28, 2024
 ##
 ##   This file is part of Logtalk <https://logtalk.org/>
 ##   Copyright 2024 Paulo Moura <pmoura@logtalk.org>
@@ -28,7 +28,8 @@
 cd "$(dirname "$0")" || exit 1
 
 TAG=$(cat ../../../VERSION.txt | sed -e 's/-stable$//' | sed -e 's/\.//g')
+TEXINFO=$(cat ../../../VERSION.txt | sed -e 's/-stable$//')
 VERSION=lgt${TAG}stable
-docker build --build-arg LOGTALK_VERSION=${VERSION} -t logtalk/logtalk3-portable -t logtalk/logtalk3-portable:${TAG} .
+docker build --build-arg LOGTALK_VERSION=${VERSION} --build-arg LOGTALK_TEXINFO_VERSION=${TEXINFO} -t logtalk/logtalk3-portable -t logtalk/logtalk3-portable:${TAG} .
 docker push logtalk/logtalk3-portable
 docker push logtalk/logtalk3-portable:${TAG}